Removed EvenActionButton changes
Removed EvenActionButton changes since they're going to be added to another branch, seperate from this one.
